Move R4MvcPreCompileModule into R4Mvc project and clear out existing generation code
