+++ src/R4MvcHostApp/Compiler/Preprocess/R4MvcCompiler.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using R4Mvc;
+
+namespace R4MvcHostApp.Compiler.Preprocess
+{
+    public class R4MvcCompiler : R4MvcPreCompileModule
+    {
+        public R4MvcCompiler(IServiceProvider serviceProvider) : base(serviceProvider)
+        {
+            
+        }
+    }
+}
